Move versioneer into _setup
